kitten-scientists: add trading support
This patch updates the kitten scientists to include support for automatic trading. Since I was able to locate trade buttons in the game API, it does not require switching to the trade tab to run. I also fixed up hunting so that parchment is a luxury good and honors a limit for furs.