buildings: invert require value when it's set to 'false'
The required resource may be set to false or undefined in order to
indicate that no resource is required. In this case, the old code set
require to true and tested against true. It seems easier to set it to
false and then test for that. In addition, explicitely set it to false, so
that we don't mess up the === strict comparison.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>