buildManager: use internal buttons instead of loaded page
This enables building even when it's not the bonfireTab loaded. Do this by
ensuring we render() the bonfireTab first time we need it. Then, we can
just search the buttons list and locate the item and use the onClick()
function. This is much cleaner than having to be loaded on the bonfire tab
directly. This is similar to how automated trading is implemented.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com> (reverted from commit 973e847d765dc1f7413658f96751cfd31ca2e93d)