kitten-scientists: trade only as much as we can hold
This patch modifies trading so that it simply searches through the race's
'sell' list and ensures that we never trade more than we can fit. However,
in order to prevent it from not trading when *any* item it produces is
capped, only prevent it from trading when all its resources are capped.
This does mean we lose some extra resources, but it means we don't starve
trading completely. This also ensures that we only trade as much as we
actually can fit. That is, if we can only store 1000 titanium, we only
trade the number of times that will *at best* give us less than that
titanium. This prevents wasting gold and catpower and the main resource
for trades which are essentially value-less.

This does reduce the amount of blueprints we get (not sure if all races
give blueprints?)

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>