trading: spread trading out over all races
This patch makes is so that we first check which races are eligble to
trade currently. Then, we only allow each race to trade at most 1/nth of
the total trades we can perform with catpower and gold this round. This
should help prevent starvation of races if you enable lots of different
races at once.

It's not necessarily ideal, because maybe you want to prioritize other
races. The best way to do that is just disable races you don't want or put
them in separate seasons. This overall change should be better generally.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>