gameLog: fix display issues with width and padding
Reduce the padding on top of the element by 10px, and increase the width
to maximum of our space. This helps get rid of ~100px of wasted space on
the right margin, which was incredibly frustrating when you saw messages
split over two lines that could fit on one. The parent div still has about
15 pixels at the right end that looks just fine.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>