options: have labels show items instead of clicking checkbox
I think I am not alone in thinking that clicking the label will show
items. Since showing and hiding labels is definitely something we will be
doing more and more of as options get more complicated, this patch makes
it so that labels will be the same as "items". Possibly we can remove the
items button entirely then, but I'm not sold on that yet.

However, it is really frustrating when you meant to make items appear and
you instead toggled the checkbox, so modify labels to be divs instead,
which will get function of showing items added if we have any.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>