buildings: ignore consumption rate
Buildings most generally want to be build even if it would consume more than
the 0.6 default amount. Add an "all" parameter, which will default to undefined
(false-y) and only set it to true for buildings. This should enable buildings
even if it would cost more than 60% of the available resources.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>